- remove separation to breakfast, dinner... - removed activities

+++ lib/features/home/presentation/widgets/intake_vertical_list.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:opennutritracker/core/domain/entity/intake_entity.dart';
 import 'package:opennutritracker/core/domain/entity/tracked_day_entity.dart';
-import 'package:opennutritracker/core/presentation/widgets/copy_dialog.dart';
 import 'package:opennutritracker/core/presentation/widgets/delete_all_dialog.dart';
 import 'package:opennutritracker/core/presentation/widgets/intake_card.dart';
 import 'package:opennutritracker/core/presentation/widgets/placeholder_card.dart';
@@ -9,7 +8,6 @@ import 'package:opennutritracker/core/utils/locator.dart';
 import 'package:opennutritracker/core/utils/navigation_options.dart';
 import 'package:opennutritracker/core/utils/vertical_list_popup_menu_selections.dart';
 import 'package:opennutritracker/features/add_meal/presentation/add_meal_screen.dart';
-import 'package:opennutritracker/features/add_meal/presentation/add_meal_type.dart';
 import 'package:opennutritracker/features/diary/presentation/bloc/calendar_day_bloc.dart';
 import 'package:opennutritracker/features/diary/presentation/bloc/diary_bloc.dart';
 import 'package:opennutritracker/features/home/presentation/bloc/home_bloc.dart';
@@ -20,7 +18,6 @@ class IntakeVerticalList extends StatefulWidget {
   final DateTime day;
   final String title;
   final IconData listIcon;
-  final AddMealType addMealType;
   final List<IntakeEntity> intakeList;
   final bool usesImperialUnits;
   final Function(IntakeEntity intake, TrackedDayEntity? trackedDayEntity)
@@ -28,8 +25,6 @@ class IntakeVerticalList extends StatefulWidget {
   final Function(BuildContext, IntakeEntity)? onItemLongPressedCallback;
   final Function(bool)? onItemDragCallback;
   final Function(BuildContext, IntakeEntity, bool)? onItemTappedCallback;
-  final Function(IntakeEntity intake, TrackedDayEntity? trackedDayEntity,
-      AddMealType? type)? onCopyIntakeCallback;
   final TrackedDayEntity? trackedDayEntity;
 
   const IntakeVerticalList({
@@ -37,14 +32,12 @@ class IntakeVerticalList extends StatefulWidget {
     required this.day,
     required this.title,
     required this.listIcon,
-    required this.addMealType,
     required this.intakeList,
     required this.usesImperialUnits,
     required this.onDeleteIntakeCallback,
     this.onItemLongPressedCallback,
     this.onItemDragCallback,
     this.onItemTappedCallback,
-    this.onCopyIntakeCallback,
     this.trackedDayEntity,
   });
 
@@ -100,39 +93,20 @@ class _IntakeVerticalListState extends State<IntakeVerticalList> {
                 PopupMenuButton<VerticalListPopupMenuSelections>(
                     onSelected:
                         (VerticalListPopupMenuSelections selection) async {
-                      switch (selection) {
-                        case VerticalListPopupMenuSelections.onCopy:
-                          const copyDialog = CopyDialog();
-                          final selectedMealType =
-                              await showDialog<AddMealType>(
-                                  context: context,
-                                  builder: (context) => copyDialog);
-                          if (selectedMealType != null) {
-                            for (IntakeEntity intake in widget.intakeList) {
-                              widget.onCopyIntakeCallback!(
-                                  intake, null, selectedMealType);
-                            }
+                      if (selection == VerticalListPopupMenuSelections.onDelete) {
+                        final shouldDeleteIntakes = await showDialog<bool>(
+                            context: context,
+                            builder: (context) => const DeleteAllDialog());
+                        if (shouldDeleteIntakes != null) {
+                          for (IntakeEntity intake in widget.intakeList) {
+                            widget.onDeleteIntakeCallback(
+                                intake, widget.trackedDayEntity);
                           }
-                          break;
-                        case VerticalListPopupMenuSelections.onDelete:
-                          final shouldDeleteIntakes = await showDialog<bool>(
-                              context: context,
-                              builder: (context) => const DeleteAllDialog());
-                          if (shouldDeleteIntakes != null) {
-                            for (IntakeEntity intake in widget.intakeList) {
-                              widget.onDeleteIntakeCallback(
-                                  intake, widget.trackedDayEntity);
-                            }
-                            break;
-                          }
+                        }
                       }
                     },
                     itemBuilder: (BuildContext context) =>
                         <PopupMenuEntry<VerticalListPopupMenuSelections>>[
-                          if (widget.onCopyIntakeCallback != null)
-                            PopupMenuItem<VerticalListPopupMenuSelections>(
-                                value: VerticalListPopupMenuSelections.onCopy,
-                                child: Text(S.of(context).dialogCopyLabel)),
                           PopupMenuItem<VerticalListPopupMenuSelections>(
                               value: VerticalListPopupMenuSelections.onDelete,
                               child: Text(S.of(context).deleteAllLabel)),
@@ -219,12 +193,12 @@ class _IntakeVerticalListState extends State<IntakeVerticalList> {
 
   void _onPlaceholderCardTapped(BuildContext context) {
     Navigator.pushNamed(context, NavigationOptions.addMealRoute,
-        arguments: AddMealScreenArguments(widget.addMealType, widget.day));
+        arguments: AddMealScreenArguments(widget.day));
   }
 
   void _onItemDropped(IntakeEntity entity) {
     _mealDetailBloc.addIntake(context, entity.unit, entity.amount.toString(),
-        widget.addMealType.getIntakeType(), entity.meal, entity.dateTime);
+        entity.meal, entity.dateTime);
     _homeBloc.deleteIntakeItem(entity);
 
     // Refresh Home Page
